Linux build script returns value indicating success or failure.

+++ build.sh
@@ -11,14 +11,21 @@ fi
 # if [ ] will build for Linux x64, if [ 1 ] builds for Linux arm64
 if [ ]; then
   dotnet publish -c Release -r linux-arm64 --self-contained=false "-p:PublishSingleFile=true" ./Recepten.sln -v n
-  cp ./bin/Release/net8.0/linux-arm64/publish/appsettings.Release.json ./bin/Release/net8.0/linux-arm64/publish/appsettings.json
-  rm ./bin/Release/net8.0/linux-arm64/publish/appsettings.*.json
-  rm ./bin/Release/net8.0/linux-arm64/publish/Dockerfile
+  if [ "$?" == "0" ]; then
+    rm ./bin/Release/net8.0/linux-arm64/publish/appsettings.*.json
+  else
+    echo "Build for Linux x64 failed"
+    return 1
+  fi
   echo "Linux arm64"
 else
   dotnet publish -c Release -r linux-x64 --self-contained=false "-p:PublishSingleFile=true" ./Recepten.sln -v n
-  cp ./bin/Release/net8.0/linux-x64/publish/appsettings.Release.json ./bin/Release/net8.0/linux-x64/publish/appsettings.json
-  rm ./bin/Release/net8.0/linux-x64/publish/appsettings.*.json
-  rm ./bin/Release/net8.0/linux-x64/publish/Dockerfile
+  if [ "$?" == "0" ]; then
+    cp ./bin/Release/net8.0/linux-x64/publish/appsettings.Release.json ./bin/Release/net8.0/linux-x64/publish/appsettings.json
+    rm ./bin/Release/net8.0/linux-x64/publish/appsettings.*.json
+  else
+    echo "Build for Linux x64 failed"
+    return 1
+  fi
   echo "Linux x64"
 fi
